MMth library functionality: segments intersections, analytical monotonic function solver, monotonic functions intersections solver

+++ src/main/scala/lib/traits/Locatable.scala
@@ -1,10 +1,10 @@
 package lib.traits
 
-import lib.Point
+import lib.math.Point
 
 trait Locatable {
-  var x: Int = 0
-  var y: Int = 0
+  var x: Double = 0
+  var y: Double = 0
   def location_=(p: Point): Unit = {
     x = p.x
     y = p.y
